requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Mail==0.9.1
Flask-Cors==4.0.0
Flask-Limiter==3.5.0
itsdangerous==2.1.2
python-dotenv==1.0.0
Werkzeug==3.0.1
Flask-JWT-Extended==4.6.0
